build_perf: add index page
A simple "welcome" page listing the tester hosts for which test results
are available.

The base html table is cannibalized and modified from the charts app.
One big difference in the base template is that bootstrap and jQuery
(the only 3rd party components used in build_perf, so far) are loaded
over the network from a CDN, instead of using the version stored in our
Git repository.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>